[#128] 백업 데이터 로딩 DB쿼리로 변경 (#129)
* refactor: 직원 수 추이 조회 API 파라미터 방식 변경 및 리턴 타입 수정

- 컨트롤러에서 EmployeeSearchCondition, Pageable 제거
- @RequestParam으로 from, to, unit 직접 받도록 수정
- 기본 unit 값은 "month"로 설정
- 서비스 리턴 타입 Page -> List로 변경
- 응답을 ResponseEntity<List<EmployeeTrendDto>>로 반환

* refactor: 인메모리 서칭 -> DB 필터링으로 변경

- 검색 처리 방식을 리팩토링하여 인메모리에서 서칭하던 로직을 DB에서 필터링하는 방식으로 변경.
- DB에서 직접 필터링 조건을 적용하고, 페이징 처리와 정렬 조건을 효율적으로 개선.
- `searchBackups` 메서드에서 `idAfter`와 `sortField`에 따른 조건을 DB 쿼리로 처리하도록 수정.

* chore: 프론트 js 코드 개행

+++ src/main/java/com/hrbank/dto/backup/CursorPageResponseBackupDto.java
@@ -7,7 +7,7 @@ public record CursorPageResponseBackupDto(
     String nextCursor,
     Long nextIdAfter,
     int size,
-    int totalElements,
+    Long totalElements,
     boolean hasNext
 ) {
